requirements.txt
typing
torch
matplotlib
matplotlib
opencv-python
pandas
argparse
Pillow
helper
numpy
tqdm
pathlib
torch
torchvision
sklearn
torchgeometry
loss
torchmetrics
trainer
analysis
segmentation_models_pytorch
simplejson
torchsummary
tensorboard